added a bunch of things and achieved 100 percent coverage, and updated docs
