+++ pyproject.toml
@@ -154,3 +154,4 @@ isort.required-imports = ["from __future__ import annotations"]
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = ["T20"]
 "noxfile.py" = ["T20"]
+"docs/**" = ["T20", "E702", "NPY002"]
